Change file-handling logic in Executor
Now in Executor, files are only opened in `train` and `test` and closed
afterwards. Originally files were opened at the point of construction,
and closed only when the program exits, using an `atexit` hook. This
keeps the Executor object alive and creates a memory leak.

Here we keep the change, but handles the logic more carefully. A few
corner cases were also considered, and we also open files in
`write_log`. This allows writing logs after construction and before
training, while the original implementation simply does nothing, which
is confusing for users.