Fix Executor bug where metrics are not evaluated
When Executor uses a SimpleMetric that is not logged, the metric is
simply not evaluated. This is problematic for metrics with side effects,
e.g. FileWriterMetric used in the BERT examples.
We now explicitly evaluate all metrics at the end of each epoch, and the
end of validation and testing.