Add finalize() method in Metric
The current way of writing predictions to file is through hacking the
metric. We previously had a `FileWriterMetric` in the BERT example,
we've now moved it to the library. Since such "metrics" are not actual
meaningful metrics, we allow it to return `None` to indicate that it
should not be included in logging.

Such metrics need to be called after the end of an epoch, validation, or
testing. Previously we used `value()` for that, but it could lead to
unintended calls or missed calls. So we introduce a `finalize()` method
specifically for this use case. The default finalizing behavior is
simply no-op.