Fix a couple of Executor bugs (#320)
* Fix: incorrect path when loading from external folder

* Change file-handling logic in Executor

Now in Executor, files are only opened in `train` and `test` and closed
afterwards. Originally files were opened at the point of construction,
and closed only when the program exits, using an `atexit` hook. This
keeps the Executor object alive and creates a memory leak.

Here we keep the change, but handles the logic more carefully. A few
corner cases were also considered, and we also open files in
`write_log`. This allows writing logs after construction and before
training, while the original implementation simply does nothing, which
is confusing for users.

* Fix Executor bug where metrics are not evaluated

When Executor uses a SimpleMetric that is not logged, the metric is
simply not evaluated. This is problematic for metrics with side effects,
e.g. FileWriterMetric used in the BERT examples.
We now explicitly evaluate all metrics at the end of each epoch, and the
end of validation and testing.

* Fix linting stuff

* Add finalize() method in Metric

The current way of writing predictions to file is through hacking the
metric. We previously had a `FileWriterMetric` in the BERT example,
we've now moved it to the library. Since such "metrics" are not actual
meaningful metrics, we allow it to return `None` to indicate that it
should not be included in logging.

Such metrics need to be called after the end of an epoch, validation, or
testing. Previously we used `value()` for that, but it could lead to
unintended calls or missed calls. So we introduce a `finalize()` method
specifically for this use case. The default finalizing behavior is
simply no-op.

* Fix stupid linting problem

* Remove unused imports